# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: EthixAI API
  description: |
    Ethical AI Governance Platform API - Real-time bias detection, SHAP explanations,
    and comprehensive monitoring for financial decision-making systems.
  version: '1.0.0'
  contact:
    name: EthixAI Support
    email: support@ethixai.com
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.ethixai.com
    description: Production
...
paths:
...
  /api/audit/logs:
    get:
      tags: [Audit]
      summary: Query audit logs
      operationId: queryAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - name: model_id
          in: query
          schema:
            type: string
        - name: event_type
          in: query
          schema:
            type: string
        - name: actor
          in: query
          schema:
            type: string
        - name: result
          in: query
          schema:
            type: string
        - name: compliance_status
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          schema:
            type: string
            default: '-timestamp'
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/audit/logs/{model_id}/trail:
    get:
      tags: [Audit]
      summary: Get audit trail for specific model
      operationId: getModelAuditTrail
      security:
        - bearerAuth: []
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
        - name: model_version
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Model audit trail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'

  # ==================== POLICIES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuditLog:
      type: object
      properties:
        _id:
          type: string
        model_id:
          type: string
        event_type:
          type: string
        actor:
          type: string
        result:
          type: string
        compliance_status:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time

    # ==================== POLICY SCHEMAS ====================
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_pages:
          type: integer
        total_records:
          type: integer
        per_page:
          type: integer
...